Add files via upload
An application that loads a bmp image and uses signal processing approach to resample the image.  class structure was developed to load, store and do simple operations like shirt, rotation and scaling. the resampling done with a convolutional kernel that is constructed as a sinc function with weighting window and the size of the convolutional filter is depending on the resampling ratio.